Sorting out the retry logic and eventing for messages.

+++ DCCPanelController/Clients/Helpers/EntryValidation.cs
@@ -36,7 +36,7 @@ public class FastClockValidationBehavior : Behavior<Entry> {
     private void OnEntryTextChanged(object? sender, TextChangedEventArgs args) {
         if (sender is Entry entry) {
             if (!string.IsNullOrEmpty(args.NewTextValue)) {
-                if (int.TryParse(args.NewTextValue, out var result)) {
+                if (double.TryParse(args.NewTextValue, out var result)) {
                     entry.TextColor = result is>= 0 and <= 12 ? Colors.Black : Colors.Red;
                 }
             }
